Revert "[NPM][USM] Move probes on sockfd_lookup_light to USM (#21259)" (#22621)
This reverts commit 18169a18e5e8dacffca8a98e5e081c21607e2aa1.
