+++ pkg/network/ebpf/c/prebuilt/usm.c
@@ -11,7 +11,6 @@
 #include "protocols/http2/decoding.h"
 #include "protocols/http2/decoding-tls.h"
 #include "protocols/kafka/kafka-parsing.h"
-#include "protocols/sockfd-probes.h"
 #include "protocols/tls/java/erpc_dispatcher.h"
 #include "protocols/tls/java/erpc_handlers.h"
 #include "protocols/tls/https.h"
