+++ pkg/network/ebpf/c/tracer.c
@@ -11,13 +11,13 @@
 #include "prebuilt/offsets.h"
 #endif
 #include "skb.h"
+#include "sockfd.h"
 #include "tracer/bind.h"
 #include "tracer/events.h"
 #include "tracer/maps.h"
 #include "tracer/port.h"
 #include "tracer/tcp_recv.h"
 #include "protocols/classification/protocol-classification.h"
-#include "protocols/sockfd.h"
 
 SEC("socket/classifier_entry")
 int socket__classifier_entry(struct __sk_buff *skb) {
@@ -1017,6 +1017,90 @@ int kretprobe__inet6_bind(struct pt_regs *ctx) {
     return sys_exit_bind(ret);
 }
 
+SEC("kprobe/sockfd_lookup_light")
+int kprobe__sockfd_lookup_light(struct pt_regs *ctx) {
+    int sockfd = (int)PT_REGS_PARM1(ctx);
+    u64 pid_tgid = bpf_get_current_pid_tgid();
+
+    // Check if have already a map entry for this pid_fd_t
+    // TODO: This lookup eliminates *4* map operations for existing entries
+    // but can reduce the accuracy of programs relying on socket FDs for
+    // processes with a lot of FD churn
+    pid_fd_t key = {
+        .pid = pid_tgid >> 32,
+        .fd = sockfd,
+    };
+    struct sock **sock = bpf_map_lookup_elem(&sock_by_pid_fd, &key);
+    if (sock != NULL) {
+        return 0;
+    }
+
+    bpf_map_update_with_telemetry(sockfd_lookup_args, &pid_tgid, &sockfd, BPF_ANY);
+    return 0;
+}
+
+static __always_inline const struct proto_ops * socket_proto_ops(struct socket *sock) {
+    const struct proto_ops *proto_ops = NULL;
+#ifdef COMPILE_PREBUILT
+    // (struct socket).ops is always directly after (struct socket).sk,
+    // which is a pointer.
+    u64 ops_offset = offset_socket_sk() + sizeof(void *);
+    bpf_probe_read_kernel_with_telemetry(&proto_ops, sizeof(proto_ops), (char*)sock + ops_offset);
+#elif defined(COMPILE_RUNTIME) || defined(COMPILE_CORE)
+    BPF_CORE_READ_INTO(&proto_ops, sock, ops);
+#endif
+
+    return proto_ops;
+}
+
+// this kretprobe is essentially creating:
+// * an index of pid_fd_t to a struct sock*;
+// * an index of struct sock* to pid_fd_t;
+SEC("kretprobe/sockfd_lookup_light")
+int kretprobe__sockfd_lookup_light(struct pt_regs *ctx) {
+    u64 pid_tgid = bpf_get_current_pid_tgid();
+    int *sockfd = bpf_map_lookup_elem(&sockfd_lookup_args, &pid_tgid);
+    if (sockfd == NULL) {
+        return 0;
+    }
+
+    // For now let's only store information for TCP sockets
+    struct socket *socket = (struct socket *)PT_REGS_RC(ctx);
+    if (!socket)
+        goto cleanup;
+
+    enum sock_type sock_type = 0;
+    bpf_probe_read_kernel_with_telemetry(&sock_type, sizeof(short), &socket->type);
+
+    const struct proto_ops *proto_ops = socket_proto_ops(socket);
+    if (!proto_ops) {
+        goto cleanup;
+    }
+
+    int family = 0;
+    bpf_probe_read_kernel_with_telemetry(&family, sizeof(family), &proto_ops->family);
+    if (sock_type != SOCK_STREAM || !(family == AF_INET || family == AF_INET6)) {
+        goto cleanup;
+    }
+
+    // Retrieve struct sock* pointer from struct socket*
+    struct sock *sock = socket_sk(socket);
+    if (!sock) {
+        goto cleanup;
+    }
+
+    pid_fd_t pid_fd = {
+        .pid = pid_tgid >> 32,
+        .fd = (*sockfd),
+    };
+
+    // These entries are cleaned up by tcp_close
+    bpf_map_update_with_telemetry(pid_fd_by_sock, &sock, &pid_fd, BPF_ANY);
+    bpf_map_update_with_telemetry(sock_by_pid_fd, &pid_fd, &sock, BPF_ANY);
+cleanup:
+    bpf_map_delete_elem(&sockfd_lookup_args, &pid_tgid);
+    return 0;
+}
 // Represents the parameters being passed to the tracepoint net/net_dev_queue
 struct net_dev_queue_ctx {
     u64 unused;
