+++ pkg/network/tracer/connection/kprobe/config.go
@@ -67,6 +67,12 @@ func enabledProbes(c *config.Config, runtimeTracer, coreTracer bool) (map[probes
 		// runtime compiled implementation
 		enableProbe(enabled, selectVersionBasedProbe(runtimeTracer || coreTracer, kv, probes.TCPRetransmit, probes.TCPRetransmitPre470, kv470))
 		enableProbe(enabled, probes.TCPRetransmitRet)
+
+		missing, err := ebpf.VerifyKernelFuncs("sockfd_lookup_light")
+		if err == nil && len(missing) == 0 {
+			enableProbe(enabled, probes.SockFDLookup)
+			enableProbe(enabled, probes.SockFDLookupRet)
+		}
 	}
 
 	if c.CollectUDPv4Conns {
