+++ pkg/network/usm/ebpf_main.go
@@ -20,6 +20,8 @@ import (
 	"github.com/davecgh/go-spew/spew"
 	"golang.org/x/sys/unix"
 
+	manager "github.com/DataDog/ebpf-manager"
+
 	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/ebpf/bytecode"
 	"github.com/DataDog/datadog-agent/pkg/network"
@@ -35,7 +37,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/usm/buildmode"
 	"github.com/DataDog/datadog-agent/pkg/network/usm/utils"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	manager "github.com/DataDog/ebpf-manager"
 )
 
 var (
@@ -110,26 +111,6 @@ func newEBPFProgram(c *config.Config, sockFD, connectionProtocolMap *ebpf.Map, b
 				},
 			},
 		},
-	}
-
-	if c.CollectTCPv4Conns || c.CollectTCPv6Conns {
-		missing, err := ddebpf.VerifyKernelFuncs("sockfd_lookup_light")
-		if err == nil && len(missing) == 0 {
-			mgr.Probes = append(mgr.Probes, []*manager.Probe{
-				{
-					ProbeIdentificationPair: manager.ProbeIdentificationPair{
-						EBPFFuncName: probes.SockFDLookup,
-						UID:          probeUID,
-					},
-				},
-				{
-					ProbeIdentificationPair: manager.ProbeIdentificationPair{
-						EBPFFuncName: probes.SockFDLookupRet,
-						UID:          probeUID,
-					},
-				},
-			}...)
-		}
 	}
 
 	program := &ebpfProgram{
@@ -379,8 +360,25 @@ func (e *ebpfProgram) init(buf bytecode.AssetReader, options manager.Options) er
 		manager.ConstantEditor{Name: "ephemeral_range_begin", Value: uint64(begin)},
 		manager.ConstantEditor{Name: "ephemeral_range_end", Value: uint64(end)})
 
-	for _, p := range e.Manager.Probes {
-		options.ActivatedProbes = append(options.ActivatedProbes, &manager.ProbeSelector{ProbeIdentificationPair: p.ProbeIdentificationPair})
+	options.ActivatedProbes = []manager.ProbesSelector{
+		&manager.ProbeSelector{
+			ProbeIdentificationPair: manager.ProbeIdentificationPair{
+				EBPFFuncName: protocolDispatcherSocketFilterFunction,
+				UID:          probeUID,
+			},
+		},
+		&manager.ProbeSelector{
+			ProbeIdentificationPair: manager.ProbeIdentificationPair{
+				EBPFFuncName: "kprobe__tcp_sendmsg",
+				UID:          probeUID,
+			},
+		},
+		&manager.ProbeSelector{
+			ProbeIdentificationPair: manager.ProbeIdentificationPair{
+				EBPFFuncName: "tracepoint__net__netif_receive_skb",
+				UID:          probeUID,
+			},
+		},
 	}
 
 	// Some parts of USM (https capturing, and part of the classification) use `read_conn_tuple`, and has some if
